Initial prototype for Downscaling as needed.
We fallback to 3D composition in case layer cannot be scanned out
directly or we don't have enough planes for all layers. In case
we fall back to 3D, we now check if we can actually downscale
the width by a given scaling factor and upscale it in display
side. We do this for offscreen surface layers only and not for
individual layer. Currently, we try to downscale only
width and not height. We ignore this in case either width or
height is less than 500 pixels. We also have a flag ENABLE_DOWNSCALING
to disable this if in case we see regressions to be fixed before
enabling this on stable.

Jira: None.
Test: No new artifacts.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>